Rename and move qlogging.hpp to common/QLogging.hpp
This is to fit the Chatterino filename standard better
