+++ src/common/Args.cpp
@@ -7,7 +7,7 @@
 #include <QJsonDocument>
 #include <QJsonObject>
 #include <QStringList>
-#include "qlogging.hpp"
+#include "common/QLogging.hpp"
 
 namespace chatterino {
