+++ src/common/QLogging.hpp
