+++ src/providers/bttv/BttvEmotes.cpp
@@ -5,12 +5,12 @@
 
 #include "common/Common.hpp"
 #include "common/NetworkRequest.hpp"
+#include "common/QLogging.hpp"
 #include "messages/Emote.hpp"
 #include "messages/Image.hpp"
 #include "messages/ImageSet.hpp"
 #include "messages/MessageBuilder.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
-#include "qlogging.hpp"
 
 namespace chatterino {
 namespace {
