+++ src/providers/twitch/PubsubHelpers.hpp
@@ -3,7 +3,7 @@
 #include <boost/asio.hpp>
 #include <boost/asio/steady_timer.hpp>
 #include <memory>
-#include "qlogging.hpp"
+#include "common/QLogging.hpp"
 #include "util/RapidjsonHelpers.hpp"
 
 namespace chatterino {
