+++ src/providers/twitch/TwitchIrcServer.cpp
@@ -6,6 +6,7 @@
 #include "Application.hpp"
 #include "common/Common.hpp"
 #include "common/Env.hpp"
+#include "common/QLogging.hpp"
 #include "controllers/accounts/AccountController.hpp"
 #include "messages/Message.hpp"
 #include "messages/MessageBuilder.hpp"
@@ -14,7 +15,6 @@
 #include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
 #include "providers/twitch/TwitchHelpers.hpp"
-#include "qlogging.hpp"
 #include "util/PostToThread.hpp"
 
 // using namespace Communi;
